Refactor MCP tools into mirrored hierarchy (#97)
