+++ Sources/kaiten-mcp/Tools/ToolCatalog.swift
@@ -1,0 +1,21 @@
+import MCP
+
+let allTools: [Tool] =
+  cardsTools
+  + cardCommentsTools
+  + cardChecklistsTools
+  + cardMembersTools
+  + cardTagsTools
+  + cardChildrenTools
+  + cardBlockersTools
+  + cardExternalLinksTools
+  + cardMetadataTools
+  + boardsTools
+  + columnsTools
+  + subcolumnsTools
+  + lanesTools
+  + spacesTools
+  + customPropertiesTools
+  + usersTools
+  + sprintsTools
+  + settingsTools
